tools/tutorials/artemis/requirements.txt
python-socketio==1.7.6
netaddr==0.7.19
Flask==0.12.2